+++ ocean-components/src/main/java/br/com/useblu/oceands/components/compose/OceanButton.kt
@@ -26,11 +26,11 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
 import br.com.useblu.oceands.R
-import br.com.useblu.oceands.ui.FontFamilyBaseBold
-import br.com.useblu.oceands.ui.OceanButtonColors
-import br.com.useblu.oceands.ui.OceanColors
-import br.com.useblu.oceands.ui.OceanFontSize
-import br.com.useblu.oceands.ui.OceanSpacing
+import br.com.useblu.oceands.ui.compose.FontFamilyBaseBold
+import br.com.useblu.oceands.ui.compose.OceanButtonColors
+import br.com.useblu.oceands.ui.compose.OceanColors
+import br.com.useblu.oceands.ui.compose.OceanFontSize
+import br.com.useblu.oceands.ui.compose.OceanSpacing
 
 @Preview(widthDp = 480)
 @Composable
