+++ ocean-components/src/main/java/br/com/useblu/oceands/ui/compose/OceanFontFamily.kt
@@ -1,4 +1,4 @@
-package br.com.useblu.oceands.ui
+package br.com.useblu.oceands.ui.compose
 
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
