+++ ocean-components/src/main/java/br/com/useblu/oceands/ui/compose/OceanSpacing.kt
@@ -1,4 +1,4 @@
-package br.com.useblu.oceands.ui
+package br.com.useblu.oceands.ui.compose
 
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
